Move cats-free and cats-state in to cats-core
