+++ core/src/main/scala/cats/free/Coyoneda.scala
