+++ core/src/main/scala/cats/free/Free.scala
