+++ core/src/main/scala/cats/free/package.scala
