+++ core/src/main/scala/cats/state/StateT.scala
