+++ core/src/main/scala/cats/state/package.scala
